+++ App/app/src/main/java/org/gammf/proxima/util/Role.java
@@ -1,4 +1,4 @@
-package org.gammf.proxima;
+package org.gammf.proxima.util;
 
 /**
  * Roles of the application domain.
