+++ App/app/src/main/res/layout/activity_data_printer.xml
@@ -5,7 +5,7 @@
     xmlns:tools="http://schemas.android.com/tools"
     android:layout_width="match_parent"
     android:layout_height="match_parent"
-    tools:context="org.gammf.proxima.DataPrinterActivity">
+    tools:context="org.gammf.proxima.android.DataPrinterActivity">
 
     <android.support.v7.widget.Toolbar
         android:id="@+id/toolbar"
